Add files via upload
This C++ project implements the Nim game with two modes: user vs. user and user vs. computer. Players take turns removing 1 or 2 tokens from piles of tokens. The number of piles can be customized or set to a default of 5 in user vs. user mode. The objective is to avoid taking the last token, with the game alternating turns between players or between the player and the computer. The game includes move validation to ensure players follow the rules. The player or computer that forces the opponent to take the last token wins.